ai_service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.27.0
openai==1.30.1
pdfplumber==0.11.0
Pillow==10.2.0
pymupdf>=1.26.0
